feat: move HireMe.vue to index.vue
