+++ frontend/public/sitemap.xml
@@ -8,25 +8,25 @@
 >
     <url>
         <loc>https://imadsaddik.com/</loc>
-        <lastmod>2026-06-16T20:25:49.025Z</lastmod>
+        <lastmod>2026-06-16T20:32:00.510Z</lastmod>
         <changefreq>daily</changefreq>
         <priority>1.0</priority>
     </url>
     <url>
         <loc>https://imadsaddik.com/about-me</loc>
-        <lastmod>2026-06-16T20:25:49.025Z</lastmod>
+        <lastmod>2026-06-16T20:32:00.510Z</lastmod>
         <changefreq>monthly</changefreq>
         <priority>0.8</priority>
     </url>
     <url>
         <loc>https://imadsaddik.com/hire-me</loc>
-        <lastmod>2026-06-16T20:25:49.025Z</lastmod>
+        <lastmod>2026-06-16T20:32:00.510Z</lastmod>
         <changefreq>monthly</changefreq>
         <priority>0.8</priority>
     </url>
     <url>
         <loc>https://imadsaddik.com/resume</loc>
-        <lastmod>2026-06-16T20:25:49.025Z</lastmod>
+        <lastmod>2026-06-16T20:32:00.510Z</lastmod>
         <changefreq>yearly</changefreq>
         <priority>0.5</priority>
     </url>
